"Connect Todo list to live database via mongo"

+++ app.js
@@ -2,6 +2,8 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const app = express();
 const mongoose = require('mongoose');
+const _ = require('lodash');
+const uri = require('./secrets/mongoURI.js');
 
 const port = 3000;
 const date = require(__dirname + '/date.js');
@@ -11,7 +13,7 @@ app.set('view engine', 'ejs');
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(express.static('public'));
 
-mongoose.connect('mongodb://localhost:27017/todoListDB', {
+mongoose.connect(uri.uri, {
   useNewUrlParser: true,
   useUnifiedTopology: true,
 });
@@ -21,6 +23,13 @@ const workItems = [];
 const itemSchema = new mongoose.Schema({
   name: String,
 });
+
+const listSchema = new mongoose.Schema({
+  name: String,
+  items: [itemSchema],
+});
+
+const List = mongoose.model('List', listSchema);
 
 const Item = mongoose.model('Item', itemSchema);
 
@@ -36,14 +45,17 @@ const item3 = new Item({
   name: '<-- Click this to delete an item.',
 });
 
+const defaultItems = [item1, item2, item3];
+
+let counter = 0;
+
 app.get('/', (req, res) => {
   const currentDay = date.getDate();
 
   Item.find({}, (err, items) => {
-    if (items.length === 0) {
-      item1.save();
-      item2.save();
-      item3.save();
+    if (items.length === 0 && counter === 0) {
+      counter++;
+      Item.insertMany(defaultItems);
       console.log('Saved items to database');
       res.redirect('/');
     } else {
@@ -57,31 +69,74 @@ app.get('/', (req, res) => {
 
 app.post('/', (req, res) => {
   const itemName = req.body.newItem;
+  const currentDay = date.getDate();
+  const currentTitle = req.body.button;
+
   const item = new Item({
     name: itemName,
   });
-  item.save();
-  res.redirect('/');
+
+  if (currentTitle === currentDay) {
+    item.save();
+    res.redirect('/');
+  } else {
+    List.findOne({ name: currentTitle }, (err, foundList) => {
+      foundList.items.push(item);
+      foundList.save();
+      res.redirect(`/${currentTitle}`);
+    });
+  }
 });
 
 app.post('/delete', (req, res) => {
   const checkedItemId = req.body.checkbox;
-  Item.findByIdAndRemove(checkedItemId, (err) => {
-    if (err) {
-      console.log(err);
-    } else {
-      console.log('Successfully deleted checked item');
+  const currentTitle = req.body.listName;
+  const currentDay = date.getDate();
+
+  if (currentTitle === currentDay) {
+    Item.findByIdAndRemove(checkedItemId, (err) => {
+      if (err) {
+        console.log(err);
+      } else {
+        console.log('Successfully deleted checked item');
+      }
+    });
+    res.redirect('/');
+  } else {
+    List.findOneAndUpdate(
+      { name: currentTitle },
+      { $pull: { items: { _id: checkedItemId } } },
+      (err, foundList) => {
+        if (!err) {
+          console.log('Successfully deleted checked item');
+          res.redirect(`/${currentTitle}`);
+        }
+      }
+    );
+  }
+});
+
+app.get('/:param', (req, res) => {
+  const custom = _.capitalize(req.params.param);
+
+  List.findOne({ name: custom }, (err, foundList) => {
+    if (!err) {
+      if (!foundList) {
+        const list = new List({
+          name: custom,
+          items: defaultItems,
+        });
+        list.save();
+        res.redirect(`/${custom}`);
+      } else {
+        console.log('found list');
+        res.render('list', {
+          listTitle: foundList.name,
+          newListItem: foundList.items,
+        });
+      }
     }
   });
-  res.redirect('/');
-});
-
-app.get('/work', (req, res) => {
-  res.render('list', { listTitle: 'Work', newListItem: workItems });
-});
-
-app.get('/about', (req, res) => {
-  res.render('about');
 });
 
 app.listen(port, () => {
